+++ MinervaApp/src/main/java/minerva/android/widget/dialog/walletconnect/DappSendTransactionDialog.kt
@@ -85,7 +85,7 @@ class DappSendTransactionDialog(context: Context, approve: () -> Unit, deny: () 
                 receiver.text = context.getText(R.string.allowance_receiver)
                 transactionType.text = context.getText(R.string.allowance)
                 amount.text =
-                    if (transaction.tokenTransaction.allowance == Int.InvalidValue.toBigDecimal()) context.getString(R.string.unlimited)
+                    if (transaction.tokenTransaction.allowance == Double.InvalidValue.toBigDecimal()) context.getString(R.string.unlimited)
                     else transaction.tokenTransaction.allowance?.toPlainString()
 
             }
